Improve blacklist (#41)
* Add date to blacklist show command

* Add option to disable blacklist commands

* Make commands customizable

* Refactor overwrite check

* Check coroutines

* Improve docs and errors

+++ ezcord/cogs/blacklist.py
@@ -46,6 +46,34 @@ async def view_check(interaction: discord.Interaction):
     return await _check_blacklist(interaction)
 
 
+def check_command(decorator):
+    """A decorator that only applies the slash command decorator if the command is enabled."""
+
+    def decorator_wrapper(func):
+        name = func.__name__
+        groups = ["blacklist", "leave"]
+        for group in groups:
+            name = name.replace(f"{group}_", "")
+
+        if name not in EzConfig.blacklist.disabled_commands:
+            return decorator(func)
+        return func
+
+    return decorator_wrapper
+
+
+async def check_overwrite(name, *args) -> bool:
+    """Returns True if the command has been overwritten, False otherwise."""
+    if not EzConfig.blacklist:
+        return False
+
+    if name in EzConfig.blacklist.overwrites:
+        await EzConfig.blacklist.overwrites[name](*args)
+        return True
+
+    return False
+
+
 class Blacklist(Cog, hidden=True):
     def __init__(self, bot: Bot):
         super().__init__(bot)
@@ -111,7 +139,7 @@ class Blacklist(Cog, hidden=True):
             description="Manage the blacklist",
         )
 
-    @blacklist.command(name="add", description="Add a member  blacklist")
+    @check_command(blacklist.command(name="add", description="Add a member  blacklist"))
     # @discord.option("user", description="The user to ban/unban")
     # @discord.option("reason", description="The reason for the ban", default=None)
     async def blacklist_add(
@@ -120,6 +148,9 @@ class Blacklist(Cog, hidden=True):
         user: discord.Member,
         reason: str = None,  # type: ignore
     ):
+        if await check_overwrite("add", ctx, user, reason):
+            return
+
         if user.id == ctx.user.id:
             return await emb.error(ctx, "You can't ban yourself.")
         if user.bot:
@@ -134,9 +165,14 @@ class Blacklist(Cog, hidden=True):
             ephemeral=True,
         )
 
-    @blacklist.command(name="remove", description="Remove a member from the blacklist")
+    @check_command(
+        blacklist.command(name="remove", description="Remove a member from the blacklist")
+    )
     # @discord.option("user", description="The user to ban/unban")
     async def blacklist_remove(self, ctx, user: discord.Member):
+        if await check_overwrite("remove", ctx, user):
+            return
+
         rowcount = await _db.remove_ban(user.id)
         if rowcount == 0:
             return await emb.error(ctx, "This user is not banned.")
@@ -144,19 +180,22 @@ class Blacklist(Cog, hidden=True):
             f"The user **{user}** was unbanned successfully.", ephemeral=True
         )
 
-    @blacklist.command(name="show", description="Show the bot blacklist")
-    async def show_blacklist(self, ctx):
+    @check_command(blacklist.command(name="show", description="Show the bot blacklist"))
+    async def blacklist_show(self, ctx):
+        if await check_overwrite("show", ctx):
+            return
+
         await ctx.response.defer(ephemeral=True)
         bans = await _db.get_full_bans()
         desc = ""
 
-        for user_id, reason, _ in bans:
+        for user_id, reason, dt in bans:
             if not reason:
                 reason = "No reason provided"
 
             user = await get_or_fetch_user(self.bot, user_id)
-            name = f"{user.name} ({user.id})" if user else user_id
-            desc += f"{name} - {reason}\n"
+            name = f"{user} - {user.id}" if user else user_id
+            desc += f"[{dt.date()}] {name} - {reason}\n"
 
         if not desc:
             desc = "No bans found."
@@ -164,8 +203,11 @@ class Blacklist(Cog, hidden=True):
         file = create_text_file(desc, "bans.txt")
         await ctx.followup.send(file=file, ephemeral=True)
 
-    @admin.command(description="Show all bot servers")
+    @check_command(admin.command(description="Show all bot servers"))
     async def show_servers(self, ctx):
+        if await check_overwrite("show_servers", ctx):
+            return
+
         await ctx.response.defer(ephemeral=True)
         longest_name = max([guild.name for guild in self.bot.guilds], key=len)
         sep = f"<{len(longest_name)}"
@@ -181,9 +223,12 @@ class Blacklist(Cog, hidden=True):
         file = create_text_file(desc, "guilds.txt")
         await ctx.followup.send(file=file, ephemeral=True)
 
-    @leave.command(name="server", description="Make the bot leave a server")
+    @check_command(leave.command(name="server", description="Make the bot leave a server"))
     # @discord.option("guild_id", description="Leave the server with the given ID", default=None)
-    async def leave_guild(self, ctx, guild_id: str):
+    async def leave_server(self, ctx, guild_id: str):
+        if await check_overwrite("server", ctx, guild_id):
+            return
+
         await ctx.response.defer(ephemeral=True)
         try:
             guild = await self.bot.fetch_guild(guild_id)
@@ -195,9 +240,14 @@ class Blacklist(Cog, hidden=True):
         await guild.leave()
         await ctx.response.send_message(f"I left **{guild.name}** ({guild.id})", ephemeral=True)
 
-    @leave.command(name="owner", description="Make the bot leave all guilds with a given owner")
+    @check_command(
+        leave.command(name="owner", description="Make the bot leave all guilds with a given owner")
+    )
     # @discord.option("owner_id", description="Leave all servers with the specified owner")
     async def leave_owner(self, ctx, owner: discord.User):
+        if await check_overwrite("owner", ctx, owner):
+            return
+
         await ctx.defer(ephemeral=True)
         guilds = []
         member_count = 0
